[REFACTOR] : add directory decorator and move files
