+++ app/src/main/java/com/dosopt/naverpay/ui/main/benefit/decorator/CardMenuItemDecoration.kt
@@ -1,4 +1,4 @@
-package com.dosopt.naverpay.ui.main.benefit
+package com.dosopt.naverpay.ui.main.benefit.decorator
 
 import android.content.Context
 import android.graphics.Rect
